// apiConnection/gen/models/sub-query-get.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LCT Hack Backend
 * Backend part for LCT Hack
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AdjustmentGet } from './adjustment-get';
import { ApartmentGet } from './apartment-get';
/**
 * 
 * @export
 * @interface SubQueryGet
 */
export interface SubQueryGet {
    /**
     * Уникальный идентификатор подзапроса
     * @type {string}
     * @memberof SubQueryGet
     */
    guid: string;
    /**
     * Список квартир в подзапросе
     * @type {Array<ApartmentGet>}
     * @memberof SubQueryGet
     */
    inputApartments?: Array<ApartmentGet>;
    /**
     * Эталонный объект
     * @type {ApartmentGet}
     * @memberof SubQueryGet
     */
    standartObject?: ApartmentGet;
    /**
     * Список подобранных аналогов
     * @type {Array<ApartmentGet>}
     * @memberof SubQueryGet
     */
    analogs?: Array<ApartmentGet>;
    /**
     * Список выбранных аналогов
     * @type {Array<ApartmentGet>}
     * @memberof SubQueryGet
     */
    selectedAnalogs?: Array<ApartmentGet>;
    /**
     * Список корректировок для аналогов
     * @type {Array<AdjustmentGet>}
     * @memberof SubQueryGet
     */
    adjustmentsAnalogCalculated?: Array<AdjustmentGet>;
    /**
     * Список корректировок для аналогов, исправленных пользователем
     * @type {Array<AdjustmentGet>}
     * @memberof SubQueryGet
     */
    adjustmentsAnalogUser?: Array<AdjustmentGet>;
    /**
     * Список корректировок для пула
     * @type {Array<AdjustmentGet>}
     * @memberof SubQueryGet
     */
    adjustmentsPoolCalculated?: Array<AdjustmentGet>;
    /**
     * Список корректировок для пула, исправленных пользователем
     * @type {Array<AdjustmentGet>}
     * @memberof SubQueryGet
     */
    adjustmentsPoolUser?: Array<AdjustmentGet>;
    /**
     * Список выходных квартир
     * @type {Array<ApartmentGet>}
     * @memberof SubQueryGet
     */
    outputApartments?: Array<ApartmentGet>;
}